Separate sanity checks from liquidity scripts
